# Conflicts:
#	langchain4j-pgvector/src/main/java/dev/langchain4j/store/embedding/pgvector/PgVectorEmbeddingStore.java

+++ langchain4j-oracle/src/main/java/dev/langchain4j/store/embedding/oracle/EmbeddingTable.java
@@ -1,17 +1,19 @@
 package dev.langchain4j.store.embedding.oracle;
 
+import static dev.langchain4j.internal.ValidationUtils.ensureNotNull;
+
+import dev.langchain4j.data.document.Metadata;
 import dev.langchain4j.data.embedding.Embedding;
 import dev.langchain4j.data.segment.TextSegment;
-
-import javax.sql.DataSource;
+import dev.langchain4j.store.embedding.filter.Filter;
 import java.sql.Connection;
 import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.List;
 import java.util.UUID;
-
-import static dev.langchain4j.internal.ValidationUtils.ensureNotNull;
-
+import java.util.function.BiFunction;
+import javax.sql.DataSource;
+import oracle.jdbc.OracleType;
 
 /**
  * <p>
@@ -104,14 +106,11 @@ public final class EmbeddingTable {
      * @throws SQLException If an error prevents the table from being created.
      */
     void create(DataSource dataSource) throws SQLException {
-        if (createOption == CreateOption.CREATE_NONE)
-            return;
+        if (createOption == CreateOption.CREATE_NONE) return;
 
         try (Connection connection = dataSource.getConnection();
-             Statement statement = connection.createStatement();
-        ) {
-            if (createOption == CreateOption.CREATE_OR_REPLACE)
-                statement.addBatch("DROP TABLE IF EXISTS " + name);
+                Statement statement = connection.createStatement(); ) {
+            if (createOption == CreateOption.CREATE_OR_REPLACE) statement.addBatch("DROP TABLE IF EXISTS " + name);
 
             statement.addBatch("CREATE TABLE IF NOT EXISTS " + name
                     + "(" + idColumn + " VARCHAR(36) NOT NULL, "
@@ -122,6 +121,33 @@ public final class EmbeddingTable {
 
             statement.executeBatch();
         }
+    }
+
+    /**
+     * <p>
+     * The mapping function for use with {@link SQLFilters#create(Filter, BiFunction)}. The function maps a
+     * {@link Metadata} key to a field of the JSON "metadata" column. The builtin JSON_VALUE function is used to
+     * evaluate a JSON path expression, which looks something like this: '$.key'
+     * </p><p>
+     * A RETURNING clause is used to return the JSON value as a particular SQL data type.
+     * </p><p>
+     * A NULL ON ERROR clause is used, explicitly, to return NULL in the case where the JSON object does not contain the
+     * key.
+     * </p>
+     *
+     * @param key Name of a metadata key. Not null.
+     * @param type SQL type to return the key as. Not null.
+     * @return A JSON_VALUE function call which returns the key as a SQL data type, or returns NULL if the key does not
+     * exist. The String returned by this method is not null.
+     */
+    String mapMetadataKey(String key, OracleType type) {
+        // Oracle JDBC does not implement getName() correctly for BINARY_FLOAT and BINARY_DOUBLE; It puts a space where
+        // the underscore should be.
+        String typeName = type == OracleType.BINARY_FLOAT
+                ? "BINARY_FLOAT"
+                : type == OracleType.BINARY_DOUBLE ? "BINARY_DOUBLE" : type.getName();
+
+        return "JSON_VALUE(" + metadataColumn + ", '$." + key + "' RETURNING " + typeName + " NULL ON ERROR)";
     }
 
     /**
@@ -286,6 +312,5 @@ public final class EmbeddingTable {
 
             return new EmbeddingTable(this);
         }
-
     }
 }
